+++ libc/pulp_malloc.c
@@ -23,7 +23,7 @@
 
 void *pi_l2_malloc(int size)
 {
-	return malloc(size);
+	return malloc((size_t)size);
 }
 
 void pi_l2_free(void *chunk, int size)
